Divide delete message tests into 2 collections
