+++ test/Telegram.Bot.Tests.Integ/Update Messages/DeleteMessageTests.cs
@@ -20,26 +20,6 @@ namespace Telegram.Bot.Tests.Integ.Update_Messages
             _fixture = fixture;
         }
 
-        [OrderedFact(DisplayName = FactTitles.ShouldDeleteMessage)]
-        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SendMessage)]
-        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.DeleteMessage)]
-        public async Task Should_Delete_Message()
-        {
-            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldDeleteMessage);
-
-            Message message = await BotClient.SendTextMessageAsync(
-                chatId: _fixture.SupergroupChat.Id,
-                text: "This message will be deleted shortly"
-            );
-
-            await Task.Delay(1_000);
-
-            await BotClient.DeleteMessageAsync(
-                chatId: message.Chat.Id,
-                messageId: message.MessageId
-            );
-        }
-
         [OrderedFact(DisplayName = FactTitles.ShouldDeleteMessageFromInlineQuery)]
         [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.AnswerInlineQuery)]
         public async Task Should_Delete_Message_From_InlineQuery()
@@ -62,20 +42,18 @@ namespace Telegram.Bot.Tests.Integ.Update_Messages
                 cacheTime: 0
             );
 
-            (Update MessageUpdate, Update ChosenResultUpdate) = await _fixture.UpdateReceiver.GetInlineQueryResultUpdates(MessageType.Text);
+            (Update messageUpdate, Update chosenResultUpdate) = await _fixture.UpdateReceiver.GetInlineQueryResultUpdates(MessageType.Text);
 
             await Task.Delay(1_000);
 
             await BotClient.DeleteMessageAsync(
-                chatId: MessageUpdate.Message.Chat.Id,
-                messageId: MessageUpdate.Message.MessageId
+                chatId: messageUpdate.Message.Chat.Id,
+                messageId: messageUpdate.Message.MessageId
             );
         }
 
         private static class FactTitles
         {
-            public const string ShouldDeleteMessage = "Should delete message";
-
             public const string ShouldDeleteMessageFromInlineQuery =
                 "Should delete message generated from an inline query result";
         }
